ci(): test e2e
